+++ CHANGELOG.md
@@ -2,10 +2,14 @@
 
 All notable changes to `sirius-helpers` will be documented in this file.
 
+## 1.1.2
+
+- Add toInitials() to StringHelpers
+
 ## 1.1.1
 
 - Add cent converter to NumberHelpers
-- Add setNullIfBlank to Sirius
+- Add setNullIfBlank() to Sirius
 
 ## 1.1.0